nlp/requirements.txt
python-dotenv
langchain
langchain_groq
langchain_community
langchain_core
streamlit
